refactor: Rename `lib/deployment/createAndSetUid.js`
