+++ lib/deployment/create-and-set-uid.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const createAndSetDeploymentUid = require('./createAndSetUid');
+const createAndSetDeploymentUid = require('./create-and-set-uid');
 
 describe('createAndSetDeploymentUid', () => {
   it('generates a random id and sets it on the contxt', () => {
