+++ lib/deployment/index.js
@@ -3,5 +3,5 @@
 module.exports = {
   parseDeploymentData: require('./parse'),
   saveDeployment: require('./save'),
-  createAndSetDeploymentUid: require('./createAndSetUid'),
+  createAndSetDeploymentUid: require('./create-and-set-uid'),
 };
